removing individual shape_calculator files and merging their code with the converters
