+++ onnxmltools/convert/sparkml/operator_converters/StringIndexer.py
@@ -25,3 +25,22 @@ def convert_sparkml_string_indexer(scope, operator, container):
 
 
 register_converter('pyspark.ml.feature.StringIndexerModel', convert_sparkml_string_indexer)
+
+import copy
+from ...common._registration import register_shape_calculator
+from ...common.data_types import Int64TensorType, StringTensorType
+from ...common.utils import check_input_and_output_numbers, check_input_and_output_types
+
+def calculate_sparkml_string_indexer_output_shapes(operator):
+    '''
+    This function just copy the input shape to the output because label encoder only alters input features' values, not
+    their shape.
+    '''
+    check_input_and_output_numbers(operator, output_count_range=1)
+    check_input_and_output_types(operator, good_input_types=[Int64TensorType, StringTensorType])
+
+    input_shape = copy.deepcopy(operator.inputs[0].type.shape)
+    operator.outputs[0].type = Int64TensorType(copy.deepcopy(input_shape))
+
+
+register_shape_calculator('pyspark.ml.feature.StringIndexerModel', calculate_sparkml_string_indexer_output_shapes)
