--- onnxmltools/convert/sparkml/shape_calculators/__init__.py
@@ -1,11 +1,0 @@
-# -------------------------------------------------------------------------
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# Licensed under the MIT License. See License.txt in the project root for
-# license information.
-# --------------------------------------------------------------------------
-
-# To register shape calculators for sparkml operators, import associated modules here.
-from . import StringIndexer
-from . import OneHotEncoder
-from . import LogisticRegression
-from . import VectorAssembler
